Move spectrum/datasets.py to datasets/spectrum.py
