+++ gammapy/cube/tests/test_simulate.py
@@ -5,7 +5,7 @@ from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.coordinates import SkyCoord
 from gammapy.cube import MapDatasetEventSampler, simulate_dataset
-from gammapy.datasets.tests.test_fit import get_map_dataset
+from gammapy.datasets.tests.test_map import get_map_dataset
 from gammapy.datasets import MapDataset
 from gammapy.data import GTI, Observation
 from gammapy.irf import load_cta_irfs
