+++ gammapy/datasets/__init__.py
@@ -1,2 +1,3 @@
 from .core import *
 from .map import *
+from .spectrum import *
