+++ gammapy/datasets/map.py
@@ -14,12 +14,14 @@ from gammapy.data import GTI
 from gammapy.irf import EDispKernel, EffectiveAreaTable
 from gammapy.maps import Map, MapAxis
 from gammapy.modeling import Parameters
-from gammapy.datasets import Dataset
 from gammapy.modeling.models import BackgroundModel, Models
 from gammapy.stats import cash, cash_sum_cython, wstat
 from gammapy.utils.random import get_random_state
 from gammapy.utils.scripts import make_name, make_path
 from gammapy.cube.exposure import _map_spectrum_weight
+from .core import Dataset
+from .spectrum import SpectrumDatasetOnOff, SpectrumDataset
+
 
 __all__ = ["MapDataset", "MapDatasetOnOff"]
 
@@ -805,8 +807,6 @@ class MapDataset(Dataset):
         dataset : `~gammapy.spectrum.SpectrumDataset`
             the resulting reduced dataset
         """
-        from gammapy.spectrum import SpectrumDataset
-
         kwargs = {"gti": self.gti, "name": name}
 
         if self.gti is not None:
@@ -1389,8 +1389,6 @@ class MapDatasetOnOff(MapDataset):
         dataset : `~gammapy.spectrum.SpectrumDatasetOnOff`
             the resulting reduced dataset
         """
-        from gammapy.spectrum import SpectrumDatasetOnOff
-
         dataset = super().to_spectrum_dataset(on_region, containment_correction, name)
 
         kwargs = {}
