+++ gammapy/datasets/tests/test_map.py
