+++ gammapy/spectrum/tests/test_sensitivity.py
@@ -4,12 +4,9 @@ import numpy as np
 from numpy.testing import assert_allclose
 import astropy.units as u
 from gammapy.irf import EDispKernel, EffectiveAreaTable
-from gammapy.spectrum import (
-    CountsSpectrum,
-    SensitivityEstimator,
-    SpectrumDataset,
-    SpectrumDatasetOnOff,
-)
+from gammapy.datasets import SpectrumDataset, SpectrumDatasetOnOff
+from gammapy.maps import CountsSpectrum
+from gammapy.spectrum import SensitivityEstimator
 
 
 @pytest.fixture()
